% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_distance.R
\name{track_distance}
\alias{track_distance}
\title{Track distance}
\usage{
track_distance(x, y)
}
\arguments{
\item{x}{longitude}

\item{y}{latitude}
}
\value{
numeric vector of distances between sequential pairs of x, y in metres, see Details
}
\description{
Calculate geodesic distance on longitude, latitude input vectors.
}
\details{
By convention the first value is set to \code{NA} missing value, because the distance
applies to each sequential pair of locations.

To use this on multiple track ids, use a grouped data frame with tidyverse code like
\code{data \%>\% group_by(id) \%>\% mutate(distance = track_distance(lon, lat))}
}
\examples{
track_distance(trips0$x, trips0$y)[1:10]
}
