% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traipse-package.R
\docType{package}
\name{traipse-package}
\alias{traipse}
\alias{traipse-package}
\title{traipse: Shared Tools for Tracking Data}
\description{
A collection of commonly used tools for animal movement and other tracking data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Trackage/traipse}
  \item Report bugs at \url{https://github.com/Trackage/traipse/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com} (0000-0002-2471-7511)

}
\keyword{internal}
