% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triplib-package.R
\docType{package}
\name{triplib-package}
\alias{triplib}
\alias{triplib-package}
\title{triplib: Shared Tools for Tracking Data}
\description{
A collection of commonly used tools for animal movement and other tracking data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Trackage/triplib}
  \item Report bugs at \url{https://github.com/Trackage/triplib/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com} (0000-0002-2471-7511)

}
\keyword{internal}
