% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traipse-package.R
\docType{package}
\name{traipse-package}
\alias{traipse}
\alias{traipse-package}
\title{traipse: Shared Tools for Tracking Data}
\description{
A collection of commonly used tools for animal movement and other tracking data. Variously distance, angle, bearing, distance-to, bearing-to and speed are provided for geographic data that can be used directly or within 'tidyverse' syntax. Distances and bearings are calculated using modern geodesic methods as provided by Charles F. F. Karney (2013) \doi{10.1007/s00190-012-0578-z} via the 'geographiclib' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AustralianAntarcticDivision/traipse}
  \item Report bugs at \url{https://github.com/AustralianAntarcticDivision/traipse/issues}
}

}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com} (\href{https://orcid.org/0000-0002-2471-7511}{ORCID})

}
\keyword{internal}
